hechas las 6 primeras pruebas estan en outputs

+++ Projecte/cercabinaria/CercaBinaria.cpp
@@ -23,6 +23,6 @@ bool CercaBinaria::existeix(int n){
 
 }
 
-int CercaBinaria::getComprovacions(){
+unsigned long long int CercaBinaria::getComprovacions(){
     return this->num_comprovacions;
 }
