+++ Projecte/cercabinaria/CercaBinaria.hpp
@@ -10,13 +10,13 @@ class CercaBinaria {
 
 	private:
 		vector<unsigned int> diccionari;
-        int num_comprovacions;
+        unsigned long long int num_comprovacions;
 		int binary_search(const vector<unsigned int>& v, int x, int esq, int dre);
 
 	public:
 
 		CercaBinaria(const vector<unsigned int>& v);
-        int getComprovacions();
+        unsigned long long int getComprovacions();
 		bool existeix(int num);
 
 };
